Refactor projection module into separate file
